Fix #4. Change integer requirements to be stricter.
Previously, whether a number was an integer or not was defined by
Python's `int` function. We now require an integer to have a unique
representation: 0053 isn't an integer, neither is -0. An integer is a
string that matches `^(0|-?[1-9]\d*)$`.

Compliance can be verified with the new example script echo.oil: The
user input '0053' should print out '0053', not '53'.