Fix #6
Cast filename arguments to string, and wrap the entire call to the
sub-interpreter in a giant try-catch, so that no matter what happens
there, it won't throw an error.

If there's something wrong with the filename (e.g. permission error, or
file doesn't exist), the command has simply no effect.